task 2

+++ 2-do_deploy_web_static.py
@@ -23,14 +23,21 @@ def do_deploy(archive_path):
         path = "/data/web_static/releases/"
         put(archive_path, '/tmp/')
         run('mkdir -p {}{}/'.format(path, no_ext))
-        run('tar -xzf /tmp/{} -C {}{}/'.format(file_n, path, no_ext))
+        result = run('tar -xzf /tmp/{} -C {}{}/'.format(file_n, path, no_ext))
+        if result.failed:
+            print(f"Failed to extract: {result.stderr}")
+            return False
         run('rm /tmp/{}'.format(file_n))
-        run('mv {0}{1}/web_static/* {0}{1}/'.format(path, no_ext))
+        result = run('rsync -a {0}{1}/web_static/* {0}{1}/'.format(path, no_ext))
+        if result.failed:
+            print(f"Rsync failed: {result.stderr}")
+            return False
         run('rm -rf {}{}/web_static'.format(path, no_ext))
         run('rm -rf /data/web_static/current')
         run('ln -s {}{}/ /data/web_static/current'.format(path, no_ext))
         print("Changes Deployed")
         return True
-    except:
+    except Exception as e:
+        print(f"Failed to deploy: {e}")
         return False
